Split ToWatchList into two components, one of them being a to-watch media card
