+++ components/AccountMenu/styles/ToWatchList.module.scss
@@ -1,8 +1,10 @@
 @import "../../../styles/variables.scss";
 
+$card-animation-duration: 0.2s;
 
 .towatch-section {
     width: 100%;
+    min-height: 100%;
     height: fit-content;
 
     padding-top: 1.5rem;
@@ -29,6 +31,9 @@
     font-weight: 400;
     font-size: 1.7rem;
     text-align: center;
+
+    padding-bottom: 0.3rem;
+    border-bottom: 1px solid $primary-color2;
 }
 
 .section__content {
@@ -57,7 +62,7 @@
     background-color: $dark-theme-bg-color1;
     border-radius: 15px;
 
-    transition: transform 0.2s ease-in-out;
+    transition: transform $card-animation-duration ease-in-out;
 }
 
 .card__img {
@@ -96,14 +101,17 @@
     margin: 0;
 }
 
+.card__img::before {
+    content: '';
+    width: 100%;
+    height: 100%;
+    transition: backdrop-filter $card-animation-duration linear;
+    position: absolute;
+    z-index: 1;
+}
+
 @media (hover: hover) and (pointer: fine) {
 
-    .card__img::before {
-        content: '';
-        width: 100%;
-        height: 100%;
-        transition: backdrop-filter 0.5s linear;
-    }
 
     .media__card:hover {
         transform: scale(1.05);
@@ -111,11 +119,7 @@
 
         
         .card__img::before {
-
-            position: absolute;
-            backdrop-filter: blur(1px);
-            
-            z-index: 1;
+         backdrop-filter: blur(0.3rem);
         }
 
     }
